--- multiprocessing/config.json
@@ -1,15 +1,0 @@
-{
-  "EPOCHS": 30,
-  "RNDM_LAYERS": 1,
-  "TRAIN_IMAGES": 60000,
-  "TEST_IMAGES": 10000,
-  "CHANNELS": 4,
-  "SAVE_PATH": "../results/",
-  "PREPROCESS": 1,
-  "DATA_REDUCTION": 1,
-  "REDUCTION_RATIO": 0.001,
-  "FILTER_TYPE": 0,
-  "POOLING_SIZE": 2,
-  "CLASSES": 10,
-  "BATCH_SIZE": 512
-}
